requirements.txt
boto3>=1.34.0
python-docx>=1.0.0
